Use SIGALRM on a timer to avoid having a background process.
I would like to just set a timer directly; unfortunely, those are per-process
(and you can't set them in the shell), so any process we fork off to set a
timer will only trigger a SIGALRM in that process.  Not useful.

So instead we run our own time-polling-loop and send SIGALRMs via kill.